ollama-rag/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
chainlit==1.0.400
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
fastapi==0.108.0
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.26
langchain-core==0.1.30
langchain-text-splitters==0.0.1
langsmith==0.1.22
Lazify==0.4.0
literalai==0.0.300
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
ollama==0.1.7
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
packaging==23.2
protobuf==4.25.3
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.32.0.post1
syncer==2.0.3
tenacity==8.2.3
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.10.0
uptrace==1.22.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
altair==5.2.0
asgiref==3.7.2
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
build==1.1.1
cachetools==5.3.3
chardet==5.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
coloredlogs==15.0.1
deepdiff==6.7.1
emoji==2.10.1
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
httptools==0.6.1
huggingface-hub==0.21.4
humanfriendly==10.0
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpath-python==1.0.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langdetect==1.0.9
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
nltk==3.8.1
oauthlib==3.2.2
onnxruntime==1.17.1
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-util-http==0.44b0
ordered-set==4.1.0
overrides==7.7.0
pandas==2.2.1
pillow==10.2.0
posthog==3.5.0
pulsar-client==3.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2024.1
rapidfuzz==3.6.2
referencing==0.33.0
regex==2023.12.25
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.31.1
sympy==1.12
tabulate==0.9.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.9.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.12.5
unstructured-client==0.21.1
uvloop==0.19.0
validators==0.22.0
websocket-client==1.7.0